Added the models and actually used start app
