--- hello_world/core/models.py
@@ -1,9 +1,0 @@
-from django.db import models
-
-class Locations(models.Model):
-    name = models.CharField(max_length=200)
-    latitude = models.FloatField()
-    longitude = models.FloatField()
-
-    def __str__(self):
-        return self.name
