+++ hello_world/settings.py
@@ -44,6 +44,7 @@ INSTALLED_APPS = [
     "django.contrib.messages",
     "django.contrib.staticfiles",
     "django_browser_reload",
+    "conditions.apps.ConditionsConfig",
 ]
 
 MIDDLEWARE = [
